requirements.txt
fastapi
pydantic
websocket
pygithub
pinggy
uvicorn
sqlalchemy